# Replace interpolated interpolation with cleaner method