@amatsuda 2017-10-26
@oniofchaos Could you squash your commits, resolve the conflict in CHANGELOG, then push -f please?